Add POST /api/basket endpoint
This creates a "basket" by iterating over the items in the basket and
adding them to the purchased_items table. This occurs in a transaction
so either all of none of the items are added.

There is currently no baskets table, but one should be added so that a
list of items may be linked back to the basket they were a part of.